% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meteostat_hourly_station.R
\name{get_meteostat_hourly_station}
\alias{get_meteostat_hourly_station}
\title{Fetch hourly Meteostat data for a station}
\usage{
get_meteostat_hourly_station(
  station,
  dateStartDate,
  dateEndDate,
  tz = Sys.timezone(),
  model = TRUE,
  cApiKey
)
}
\arguments{
\item{station}{- character scalar, ID of the station as given by
Meteostat's dictionary table of weather stations as from \code{RMeteostat::get_meteostat_stations()}}

\item{dateStartDate}{- Date class scalar, first date of the interval for which
the data is to be obtained}

\item{dateEndDate}{- Date class scalar, last date of the interval for which
the data is to be obtained}

\item{tz}{- timezone to return, will also include utc.  Defaults to your \code{Sys.timezone()}
otherwise use one of \code{OlsonNames()}}

\item{model}{- logical scalar. Substitute missing records with statistically optimized model data}

\item{cApiKey}{- character scalar, your Meteostat's api key}
}
\value{
A \code{data.frame} with the data fetched; the col names are self-explanatory,
See the API's documentation \url{https://dev.meteostat.net/api/stations/hourly.html#response} for more information
}
\description{
Fetch hourly Meteostat data for a Station for at most 30 day interval
}
\details{
This function fetches data made available via Meteostat's API. Please note that
due to the cap on the size of the window for which the data can be fetched,
this function can only handle requests that ask for data for up to 30 days.
See: \url{https://dev.meteostat.net/api/stations/hourly.html} for more information
}
\examples{
\dontrun{
api_key = "------my key------"
station="KMIV0"
start = as.Date("2022-05-01")
end = as.Date("2022-05-07")
df.hourly.stn <- get_meteostat_hourly_station(station, start, end, cApiKey = api_key)
}


}
