% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meteostat_stations.R
\name{get_meteostat_stations}
\alias{get_meteostat_stations}
\title{Download the full list set of stations and convert to sf}
\usage{
get_meteostat_stations()
}
\value{
sf with the following columns:
}
\description{
Download the full list of Meteostat's stations (updated daily)
}
\details{
This function downloads and parses the file with the full list of meteo stations
and returns a simple features to feed data to Meteostat's API.
See: https://dev.meteostat.net/bulk/stations.html#endpoints
}
