% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meteostat_daily_point.R
\name{get_meteostat_daily_point}
\alias{get_meteostat_daily_point}
\title{Fetch daily Meteostat data for a station}
\usage{
get_meteostat_daily_point(
  lat,
  lon,
  alt = NULL,
  dateStartDate,
  dateEndDate,
  model = TRUE,
  cApiKey
)
}
\arguments{
\item{lat}{- latitud of point}

\item{lon}{- longitude of point}

\item{alt}{- altitude of point}

\item{dateStartDate}{- Date class scalar, first date of the interval for which
the data is to be obtained}

\item{dateEndDate}{- Date class scalar, last date of the interval for which
the data is to be obtained}

\item{model}{-      Substitute missing records with statistically optimized model data}

\item{cApiKey}{- character scalar, Meteostat's key to use to fetch the data}
}
\value{
dataframe with the data fetched; the colnames are self-explanatory,
in case of doubt cf. the implementation and API's documentation:
https://dev.meteostat.net/api/stations/daily.html#response
}
\description{
Fetch daily Meteostat data for a station for at most yearly interval
}
\details{
This function fetches data made available via Meteostat's API. Please note that
due to the cap on the size of the window for which the data can be fetched,
this function can only handle requests that ask for data for up to 370 days.
For generalized version of this function capable of handling arbitrary
time intervals cf. dtGetDailyStationData
}
