% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meteostat_hourly_station.R
\name{get_meteostat_hourly_station}
\alias{get_meteostat_hourly_station}
\title{Fetch hourly Meteostat data for a station}
\usage{
get_meteostat_hourly_station(
  station,
  dateStartDate,
  dateEndDate,
  tz = Sys.timezone(),
  model = TRUE,
  cApiKey
)
}
\arguments{
\item{station}{- station id}

\item{dateStartDate}{- Date class scalar, first date of the interval for which
the data is to be obtained}

\item{dateEndDate}{- Date class scalar, last date of the interval for which
the data is to be obtained}

\item{tz}{- timezone to return, will also include utc.  defaults to Sys.timezone()}

\item{model}{- Substitute missing records with statistically optimized model data}

\item{cApiKey}{- character scalar, Meteostat's key to use to fetch the data}
}
\value{
dataframe with the data fetched; the colnames are self-explanatory,
in case of doubt cf. the implementation and API's documentation:
https://dev.meteostat.net/api/stations/daily.html#response
}
\description{
Fetch hourly Meteostat data for a Station for at most 30 day interval
}
\details{
This function fetches data made available via Meteostat's API. Please note that
due to the cap on the size of the window for which the data can be fetched,
this function can only handle requests that ask for data for up to 30 days.
}
